daachorse

Trait Serializable

pub trait Serializable: Sized {
    // Required methods
    fn serialize_to_vec(&self, dst: &mut Vec<u8>);
    fn deserialize_from_slice(src: &[u8]) -> (Self, &[u8]);
    fn serialized_bytes() -> usize;
}

Trait indicating serializability.

If the type of output value of the automaton implements this trait, the automaton can be serialized.

Required Methods

fn serialize_to_vec(&self, dst: &mut Vec<u8>)

A function called during serialization.

Arguments

dst - the destination to which the serialized data is written.

fn deserialize_from_slice(src: &[u8]) -> (Self, &[u8])

A function called during deserialization. This function must return the pair of the struct and the rest slice.

Arguments

src - the source slice containing the serialized data.

fn serialized_bytes() -> usize

Returns the size of serialized data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.